lenet for multiclass classification

+++ 3rdparty/dlib/opencv/cv_image_abstract.h
@@ -127,7 +127,7 @@ namespace dlib
                 - returns the number of columns in this image
         !*/
 
-        size_t size (
+        unsigned long size (
         ) const; 
         /*!
             ensures
